pyproject.toml
[project]
name = "jerryscanai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anomalib>=2.2.0",
    "fastapi>=0.110.0",
    "lightning>=2.6.1",
    "numpy>=2.3.5",
    "opencv-python>=4.13.0.92",
    "openvino>=2025.4.1",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.4.0",
    "streamlit>=1.32.0",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
]
